src/pyproject.toml
[tool.poetry]
name = "planeks"
version = "0.1.0"
description = ""
authors = ["cnsr <dmytrenkoyar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
celery = "^4.4.7"
django-extensions = "^3.0.8"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.6"
djangorestframework = "^3.11.1"
djangorestframework-jwt = "^1.11.0"
django-cors-headers = "^3.5.0"
redis = "^3.5.3"
faker = "^4.1.2"
django-storages = "^1.10"
boto3 = "^1.14.59"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
